* corrected from ConvertFromLatLong(...) to use double instead of float    - this is a bug that was corrected in version 4.0.3, which caused      incorrect rounding for lat/long coordinates

+++ calculation/utility/UtilityFunctions.h
@@ -7,7 +7,7 @@
 #define DEBUGLATLONG 0
 
 double 			CalcLogLikelihood(count_t n, measure_t u, count_t N, measure_t U);
-void 			ConvertFromLatLong(float Latitude, float Longitude, std::vector<double>& vCoordinates);
+void 			ConvertFromLatLong(double Latitude, double Longitude, std::vector<double>& vCoordinates);
 void 			ConvertToLatLong(float* Latitude, float* Longitude, double* pCoords);
 void 			DisplayVersion(FILE* fp, int nPos);
 double                  GetDuczmalCorrection(double dEllipseShape);
